# Improve python performance