--- a/pyconnx/connx/__init__.py
+++ b/pyconnx/connx/__init__.py
@@ -158,9 +158,7 @@ class Model(Wrapper):
         outputs = (ctypes.POINTER(Tensor._wrapped_class_) * output_count.value)()
 
         for t in input_data:
-            # XXX: Don't know why, but it's unreferenced twice
-            bindings.tensor_ref(t._wrapped_object)
-            bindings.tensor_ref(t._wrapped_object)
+            bindings.tensor_ref_child(t._wrapped_object)
 
         ret = bindings.model_run(
             self._wrapped_object,
@@ -172,17 +170,48 @@ class Model(Wrapper):
         if ret != 0:
             raise RuntimeError(f'Model run failed: {ret}')
 
+        for t in input_data:
+            # XXX: Don't know why, but it's unreferenced twice
+            # To avoid unwanted free, fix ref_count
+            if t._wrapped_object.ref_count <= 0:
+                t._wrapped_object.ref_count = 1
+            bindings.tensor_unref_child(t._wrapped_object)
+
         return [Tensor(outputs[i].contents) for i in range(output_count.value)]
 
     def benchmark(self, input_data: List[Tensor], repeat: int = 10, *, aggregate=True) -> Union[List[float], float]:
+        input_count = len(input_data)
+        inputs = (ctypes.POINTER(Tensor._wrapped_class_) * len(input_data))(
+            *[ctypes.pointer(t._wrapped_object) for t in input_data])
+        output_count = ctypes.c_uint32(self.graphs[0].contents.output_count)
+        outputs = (ctypes.POINTER(Tensor._wrapped_class_) * output_count.value)()
+
+        for t in input_data:
+            # XXX: Don't know why, but it's unreferenced twice
+            bindings.tensor_ref_child(t._wrapped_object)
 
         def func():
-            self.run(input_data)
+
+            bindings.model_run(
+                self._wrapped_object,
+                input_count,
+                inputs,
+                ctypes.byref(output_count),
+                outputs)
 
         if aggregate:
-            return timeit.timeit(func, number=repeat) / repeat
+            result = timeit.timeit(func, number=repeat) / repeat
         else:
-            return timeit.repeat(func, number=repeat)
+            result = timeit.repeat(func, repeat=repeat, number=1)
+
+        for t in input_data:
+            # XXX: Don't know why, but it's unreferenced twice
+            # To avoid unwanted free, fix ref_count
+            if t._wrapped_object.ref_count <= 0:
+                t._wrapped_object.ref_count = 1
+            bindings.tensor_unref_child(t._wrapped_object)
+
+        return result
 
     def __repr__(self):
         name = self.__class__.__name__