* Don't use clone on `Model.run`. use `ref_count` and send directly
  * Also use `child_count` to run model repeatedly on `Model.benchmark`
  * Exclude unreference time on `Model.benchmark`
* Make `Tensor.clone` faster
* Fix memory leak. call unref on `__del__`, But use a single flag to avoid double free
* Disable GC while benchmark to measure steady time

Result: (nanoseconds for total 10_000 round)
| test_name | connx(before) | connx(after) | tflite |
|:--:|--:|--:|--:|
test_add | 2,393,216,333 | 18,881,798 | 14,430,406
test_add_bcast | 2,471,384,676 | 27,010,202 | 14,642,340
test_add_uint8 | 2,359,203,716 | 18,350,124 | Missing